+++ .github/workflows/ci.yaml
@@ -21,7 +21,7 @@ jobs:
       - uses: actions-rs/toolchain@v1
         with:
           profile: minimal
-          toolchain: stable
+          toolchain: 1.66.1
           components: clippy
       - uses: Swatinem/rust-cache@v1
       - run: ./scripts/ci/lint
